pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["sarahamelo <sarah.alves.melo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
django = "^5.1.4"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
